Reorganize splitting code
