+++ tests/test_triples_factory.py
@@ -9,9 +9,8 @@ import torch
 
 from pykeen.datasets import Nations
 from pykeen.triples import TriplesFactory, TriplesNumericLiteralsFactory
-from pykeen.triples.triples_factory import (
-    INVERSE_SUFFIX, TRIPLES_DF_COLUMNS, _tf_cleanup_all, _tf_cleanup_deterministic, _tf_cleanup_randomized,
-)
+from pykeen.triples.splitting import _tf_cleanup_all, _tf_cleanup_deterministic, _tf_cleanup_randomized
+from pykeen.triples.triples_factory import INVERSE_SUFFIX, TRIPLES_DF_COLUMNS
 
 triples = np.array(
     [
